requirements.txt
openai==0.27.8
langchain==0.0.312
chromadb==0.4.14
langchainhub==0.1.13
python-dotenv==1.0.0
pdfplumber==0.10.2
tiktoken==0.5.1
streamlit==1.27.2